requirements.txt
backcall==0.2.0
coloredlogs==15.0
et-xmlfile==1.0.1
humanfriendly==9.1
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
openpyxl==3.0.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.14.0
ptyprocess==0.7.0
pygments==2.8.1
tensorboardx==2.1
toml==0.10.2
torchsummary==1.5.1
traitlets==5.0.5
wcwidth==0.2.5
torchvision>=0.5.0
natsort
data
scipy
matplotlib
chardet
tqdm